requirements.txt
altair==5.4.1
APScheduler==3.11.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
Cartopy==0.24.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==44.0.2
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
Flask==3.1.0
fonttools==4.54.1
fpdf==1.7.2
fpf==0.1.4
fsspec==0.7.4
gitdb==4.0.11
GitPython==3.1.43
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpickle==3.3.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
midi2audio==0.1.1
more-itertools==10.5.0
MouseInfo==0.1.3
music21==9.1.0
mysql-connector==2.2.9
narwhals==1.9.2
numpy==2.1.2
opencv-python==4.11.0.86
openpyxl==3.1.5
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pdf2image==1.17.0
pdfkit==1.0.0
pillow==10.4.0
playsound3==2.5.1
protobuf==5.28.2
psutil==7.0.0
pyarrow==17.0.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pybrowsers==0.6.0
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
pydyf==0.11.0
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
PyMsgBox==1.0.9
pyparsing==3.1.4
pypdf==5.4.0
pyperclip==1.9.0
pyphen==0.17.2
pyproj==3.7.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyRect==0.2.0
PyScreeze==1.0.1
pyshp==2.3.1
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytweening==1.2.0
pytz==2024.2
pywhatkit==5.4
pywin32==306
PyYAML==6.0.2
qrcode==8.1
referencing==0.35.1
reportlab==4.4.0
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
schedule==1.2.2
selenium==4.29.0
shapely==2.0.7
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
streamlit==1.39.0
svglib==1.5.1
tenacity==9.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
watchdog==5.0.3
weasyprint==65.1
webcolors==24.8.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.9.0
wikipedia==1.4.0
wsproto==1.2.0
xhtml2pdf==0.2.17
xlrd==2.0.1
yt-dlp==2024.12.13
zopfli==0.2.3.post1
